+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 
 - (**Extension**) Install extension when clicking card of uninstalled extension
+- (**Updates/History/Downloads**) Show scanlator of chapter
 
 ### Fixed